requirement.txt
Usage: pip install -r requirements.txt

# Base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.41.0
protobuf<4.21.3
tensorflow>=2.4.1  # TFLite export
tensorflow_datasets
-q tfds-nightly
git+https://github.com/malekijoo/kerod


# Logging -------------------------------------
tensorboard>=2.4.1

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Extras --------------------------------------
ipython  # interactive notebook
psutil  # system utilization
thop  # FLOPs computation